dt1 editor: subtile flags editor feature
